Disable Qt Test for win-clang-undefined.
It seems even a basic QString::toStdString() throws a
violation. I guess it is not supported. Qt definitely doesn't
advertise itself as being sanitizer clean. Maybe Qt needs
to be build with sanitizer checks, although thats usually not
required for ASAN. So for now we disable it, until we extend
sanitizers to Qt builds.